+++ Unity/Assets/_Project/Scripts/Player/MobileStats.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: d41342493c7e48fe86785776c2e845fe
+timeCreated: 1707054359
